requirements.txt
click
tqdm
matplotlib
numpy
pytorch-gpu
